# Optimize slp encode