I was investigating https://github.com/ssbc/ssb-db2/issues/346 and trying to get an overview of what the difference between box1 and box2 is and if we can do some optimizations. I found that slp encode is rather slow. This optimizes that function to do minimal allocations. 

Before and after benchmark from db2:

```
# private box2 (1 recps)
ok 15 unbox 1000 box2 msgs first run: 180.65ms
ok 16 unbox 1000 box2 msgs second run: 94.69ms
# private box2 group keys (1 key)
ok 19 unbox 1000 box2 group msgs first run: 247.77ms
ok 20 unbox 1000 box2 group msgs second run: 114.02ms

--- after optimizations

# private box2
ok 15 unbox 1000 box2 msgs first run: 161.20ms
ok 16 unbox 1000 box2 msgs second run: 83.33ms
# private box2 group keys
ok 19 unbox 1000 box2 group msgs first run: 222.86ms
ok 20 unbox 1000 box2 group msgs second run: 103.04ms
```